pyproject.toml
[project]
name = "pdf-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.29.0",
    "pillow>=11.2.1",
    "pypdf>=5.4.0",
    "pytesseract>=0.3.13",
]
